Update README.md and remove redundant file
The README file has been updated with an organized layout for better readability. A redundant README file - README (1).md - was deleted as it essentially duplicated content. Further, a useless introductory text for the SDK was removed from the README.md file. Now the readme consists of vitally important content, making it more user-friendly.